requirements.txt
feedparser
google-generativeai
supabase
python-dotenv
requests
dashscope
openai